Fix: Selectbox 아이콘에 이벤트핸들러 안달려있는 버그 수정
